feat: simplifying towards docling-parse v5
