+++ docling/backend/docling_parse_backend.py
@@ -1,155 +1,132 @@
 import logging
-import random
 from collections.abc import Iterable
 from io import BytesIO
 from pathlib import Path
-from typing import List, Optional, Union
+from typing import TYPE_CHECKING, Optional, Union
 
 import pypdfium2 as pdfium
-from docling_core.types.doc import BoundingBox, CoordOrigin, Size
-from docling_core.types.doc.page import (
-    BoundingRectangle,
-    SegmentedPdfPage,
-    TextCell,
-)
-from docling_parse.pdf_parsers import pdf_parser_v1
-from PIL import Image, ImageDraw
+from docling_core.types.doc import BoundingBox, CoordOrigin
+from docling_core.types.doc.page import SegmentedPdfPage, TextCell
+from docling_parse.pdf_parser import DoclingPdfParser, PdfDocument
+from PIL import Image
 from pypdfium2 import PdfPage
 
 from docling.backend.pdf_backend import PdfDocumentBackend, PdfPageBackend
-from docling.backend.pypdfium2_backend import get_pdf_page_geometry
-from docling.datamodel.document import InputDocument
+from docling.datamodel.backend_options import PdfBackendOptions
+from docling.datamodel.base_models import Size
+from docling.utils.locks import pypdfium2_lock
+
+if TYPE_CHECKING:
+    from docling.datamodel.document import InputDocument
 
 _log = logging.getLogger(__name__)
 
 
 class DoclingParsePageBackend(PdfPageBackend):
     def __init__(
-        self, parser: pdf_parser_v1, document_hash: str, page_no: int, page_obj: PdfPage
+        self,
+        *,
+        dp_doc: PdfDocument,
+        page_obj: PdfPage,
+        page_no: int,
+        create_words: bool = True,
+        create_textlines: bool = True,
+        keep_chars: bool = False,
+        keep_lines: bool = False,
+        keep_images: bool = True,
     ):
         self._ppage = page_obj
-        parsed_page = parser.parse_pdf_from_key_on_page(document_hash, page_no)
+        self._dp_doc = dp_doc
+        self._page_no = page_no
 
-        self.valid = "pages" in parsed_page
-        if self.valid:
-            self._dpage = parsed_page["pages"][0]
-        else:
-            _log.info(
-                f"An error occurred when loading page {page_no} of document {document_hash}."
-            )
+        self._create_words = create_words
+        self._create_textlines = create_textlines
+
+        self._keep_chars = keep_chars
+        self._keep_lines = keep_lines
+        self._keep_images = keep_images
+
+        self._dpage: Optional[SegmentedPdfPage] = None
+        self._unloaded = False
+        self.valid = (self._ppage is not None) and (self._dp_doc is not None)
+
+    def _ensure_parsed(self) -> None:
+        if self._dpage is not None:
+            return
+
+        seg_page = self._dp_doc.get_page(
+            self._page_no + 1,
+            keep_chars=self._keep_chars,
+            keep_lines=self._keep_lines,
+            keep_bitmaps=self._keep_images,
+            create_words=self._create_words,
+            create_textlines=self._create_textlines,
+            enforce_same_font=True,
+        )
+
+        # In Docling, all TextCell instances are expected with top-left origin.
+        [
+            tc.to_top_left_origin(seg_page.dimension.height)
+            for tc in seg_page.textline_cells
+        ]
+        [tc.to_top_left_origin(seg_page.dimension.height) for tc in seg_page.char_cells]
+        [tc.to_top_left_origin(seg_page.dimension.height) for tc in seg_page.word_cells]
+
+        self._dpage = seg_page
 
     def is_valid(self) -> bool:
         return self.valid
 
-    def _compute_text_cells(self) -> List[TextCell]:
-        """Compute text cells from docling-parse data."""
-        cells: List[TextCell] = []
-        cell_counter = 0
-
-        if not self.valid:
-            return cells
-
-        page_size = self.get_size()
-
-        parser_width = self._dpage["width"]
-        parser_height = self._dpage["height"]
-
-        for i in range(len(self._dpage["cells"])):
-            rect = self._dpage["cells"][i]["box"]["device"]
-            x0, y0, x1, y1 = rect
-
-            if x1 < x0:
-                x0, x1 = x1, x0
-            if y1 < y0:
-                y0, y1 = y1, y0
-
-            text_piece = self._dpage["cells"][i]["content"]["rnormalized"]
-            cells.append(
-                TextCell(
-                    index=cell_counter,
-                    text=text_piece,
-                    orig=text_piece,
-                    from_ocr=False,
-                    rect=BoundingRectangle.from_bounding_box(
-                        BoundingBox(
-                            l=x0 * page_size.width / parser_width,
-                            b=y0 * page_size.height / parser_height,
-                            r=x1 * page_size.width / parser_width,
-                            t=y1 * page_size.height / parser_height,
-                            coord_origin=CoordOrigin.BOTTOMLEFT,
-                        )
-                    ).to_top_left_origin(page_size.height),
-                )
-            )
-
-            cell_counter += 1
-
-        return cells
-
     def get_text_in_rect(self, bbox: BoundingBox) -> str:
-        if not self.valid:
-            return ""
+        self._ensure_parsed()
+        assert self._dpage is not None
+
         # Find intersecting cells on the page
         text_piece = ""
         page_size = self.get_size()
-        parser_width = self._dpage["width"]
-        parser_height = self._dpage["height"]
 
         scale = (
             1  # FIX - Replace with param in get_text_in_rect across backends (optional)
         )
 
-        for i in range(len(self._dpage["cells"])):
-            rect = self._dpage["cells"][i]["box"]["device"]
-            x0, y0, x1, y1 = rect
-            cell_bbox = BoundingBox(
-                l=x0 * scale * page_size.width / parser_width,
-                b=y0 * scale * page_size.height / parser_height,
-                r=x1 * scale * page_size.width / parser_width,
-                t=y1 * scale * page_size.height / parser_height,
-                coord_origin=CoordOrigin.BOTTOMLEFT,
-            ).to_top_left_origin(page_height=page_size.height * scale)
+        for i, cell in enumerate(self._dpage.textline_cells):
+            cell_bbox = (
+                cell.rect.to_bounding_box()
+                .to_top_left_origin(page_height=page_size.height)
+                .scaled(scale)
+            )
 
             overlap_frac = cell_bbox.intersection_over_self(bbox)
 
             if overlap_frac > 0.5:
                 if len(text_piece) > 0:
                     text_piece += " "
-                text_piece += self._dpage["cells"][i]["content"]["rnormalized"]
+                text_piece += cell.text
 
         return text_piece
 
     def get_segmented_page(self) -> Optional[SegmentedPdfPage]:
-        if not self.valid:
-            return None
-
-        text_cells = self._compute_text_cells()
-
-        # Get the PDF page geometry from pypdfium2
-        dimension = get_pdf_page_geometry(self._ppage)
-
-        # Create SegmentedPdfPage
-        return SegmentedPdfPage(
-            dimension=dimension,
-            textline_cells=text_cells,
-            char_cells=[],
-            word_cells=[],
-            has_lines=len(text_cells) > 0,
-            has_words=False,
-            has_chars=False,
-        )
+        self._ensure_parsed()
+        return self._dpage
 
     def get_text_cells(self) -> Iterable[TextCell]:
-        return self._compute_text_cells()
+        self._ensure_parsed()
+        assert self._dpage is not None
+
+        return self._dpage.textline_cells
 
     def get_bitmap_rects(self, scale: float = 1) -> Iterable[BoundingBox]:
+        self._ensure_parsed()
+        assert self._dpage is not None
+
         AREA_THRESHOLD = 0  # 32 * 32
 
-        for i in range(len(self._dpage["images"])):
-            bitmap = self._dpage["images"][i]
-            cropbox = BoundingBox.from_tuple(
-                bitmap["box"], origin=CoordOrigin.BOTTOMLEFT
-            ).to_top_left_origin(self.get_size().height)
+        images = self._dpage.bitmap_resources
+
+        for img in images:
+            cropbox = img.rect.to_bounding_box().to_top_left_origin(
+                self.get_size().height
+            )
 
             if cropbox.area() > AREA_THRESHOLD:
                 cropbox = cropbox.scaled(scale=scale)
@@ -177,54 +154,89 @@ class DoclingParsePageBackend(PdfPageBackend):
             padbox.r = page_size.width - padbox.r
             padbox.t = page_size.height - padbox.t
 
-        image = (
-            self._ppage.render(
-                scale=scale * 1.5,
-                rotation=0,  # no additional rotation
-                crop=padbox.as_tuple(),
-            )
-            .to_pil()
-            .resize(size=(round(cropbox.width * scale), round(cropbox.height * scale)))
-        )  # We resize the image from 1.5x the given scale to make it sharper.
+        with pypdfium2_lock:
+            image = (
+                self._ppage.render(
+                    scale=scale * 1.5,
+                    rotation=0,  # no additional rotation
+                    crop=padbox.as_tuple(),
+                )
+                .to_pil()
+                .resize(
+                    size=(round(cropbox.width * scale), round(cropbox.height * scale))
+                )
+            )  # We resize the image from 1.5x the given scale to make it sharper.
 
         return image
 
     def get_size(self) -> Size:
-        return Size(width=self._ppage.get_width(), height=self._ppage.get_height())
+        with pypdfium2_lock:
+            return Size(width=self._ppage.get_width(), height=self._ppage.get_height())
+
+        # TODO: Take width and height from docling-parse.
+        # return Size(
+        #    width=self._dpage.dimension.width,
+        #    height=self._dpage.dimension.height,
+        # )
 
     def unload(self):
+        if not self._unloaded and self._dp_doc is not None:
+            self._dp_doc.unload_pages((self._page_no + 1, self._page_no + 2))
+            self._unloaded = True
+
         self._ppage = None
         self._dpage = None
+        self._dp_doc = None
 
 
 class DoclingParseDocumentBackend(PdfDocumentBackend):
-    def __init__(self, in_doc: "InputDocument", path_or_stream: Union[BytesIO, Path]):
-        super().__init__(in_doc, path_or_stream)
+    def __init__(
+        self,
+        in_doc: "InputDocument",
+        path_or_stream: Union[BytesIO, Path],
+        options: PdfBackendOptions = PdfBackendOptions(),
+    ):
+        super().__init__(in_doc, path_or_stream, options)
 
-        self._pdoc = pdfium.PdfDocument(self.path_or_stream)
-        self.parser = pdf_parser_v1()
-
-        success = False
-        if isinstance(self.path_or_stream, BytesIO):
-            success = self.parser.load_document_from_bytesio(
-                self.document_hash, self.path_or_stream
-            )
-        elif isinstance(self.path_or_stream, Path):
-            success = self.parser.load_document(
-                self.document_hash, str(self.path_or_stream)
-            )
+        password = (
+            self.options.password.get_secret_value() if self.options.password else None
+        )
+        with pypdfium2_lock:
+            self._pdoc = pdfium.PdfDocument(self.path_or_stream, password=password)
+        self.parser = DoclingPdfParser(loglevel="fatal")
+        self.dp_doc: PdfDocument = self.parser.load(
+            path_or_stream=self.path_or_stream, password=password
+        )
+        success = self.dp_doc is not None
 
         if not success:
             raise RuntimeError(
-                f"docling-parse could not load document with hash {self.document_hash}."
+                f"docling-parse could not load document {self.document_hash}."
             )
 
     def page_count(self) -> int:
-        return len(self._pdoc)  # To be replaced with docling-parse API
+        # return len(self._pdoc)  # To be replaced with docling-parse API
 
-    def load_page(self, page_no: int) -> DoclingParsePageBackend:
+        len_1 = len(self._pdoc)
+        len_2 = self.dp_doc.number_of_pages()
+
+        if len_1 != len_2:
+            _log.error(f"Inconsistent number of pages: {len_1}!={len_2}")
+
+        return len_2
+
+    def load_page(
+        self, page_no: int, create_words: bool = True, create_textlines: bool = True
+    ) -> DoclingParsePageBackend:
+        with pypdfium2_lock:
+            ppage = self._pdoc[page_no]
+
         return DoclingParsePageBackend(
-            self.parser, self.document_hash, page_no, self._pdoc[page_no]
+            dp_doc=self.dp_doc,
+            page_obj=ppage,
+            page_no=page_no,
+            create_words=create_words,
+            create_textlines=create_textlines,
         )
 
     def is_valid(self) -> bool:
@@ -232,6 +244,17 @@ class DoclingParseDocumentBackend(PdfDocumentBackend):
 
     def unload(self):
         super().unload()
-        self.parser.unload_document(self.document_hash)
-        self._pdoc.close()
-        self._pdoc = None
+        # Unload docling-parse document first
+        if self.dp_doc is not None:
+            self.dp_doc.unload()
+            self.dp_doc = None
+
+        # Then close pypdfium2 document with proper locking
+        if self._pdoc is not None:
+            with pypdfium2_lock:
+                try:
+                    self._pdoc.close()
+                except Exception:
+                    # Ignore cleanup errors
+                    pass
+            self._pdoc = None
