+++ pyproject.toml
@@ -46,7 +46,7 @@ requires-python = '>=3.9,<4.0'
 dependencies = [
   'pydantic (>=2.0.0,<3.0.0)',
   'docling-core[chunking] (>=2.50.1,<3.0.0)',
-  'docling-parse (>=4.7.0,<5.0.0)',
+  'docling-parse @ git+https://github.com/docling-project/docling-parse.git@dev/prepare-for-v5',
   "docling-ibm-models>=3.9.1,<4",
   'filetype (>=1.2.0,<2.0.0)',
   'pypdfium2 (>=4.30.0,!=4.30.1,<5.0.0)',
